Application/requirements.txt
aiofiles==0.4.0
aiohttp==3.5.4
altgraph==0.16.1
async-timeout==3.0.1
attrs==19.1.0
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
cx-Freeze==5.1.1
Cython==0.28.2
decorator==4.0.11
docutils==0.14
Faker==1.0.4
future==0.17.1
httptools==0.0.13
humanfriendly==4.18
idna==2.8
idna-ssl==1.1.0
imageio==2.1.2
-e git+https://github.com/LevPasha/Instagram-API-python.git@cf0fe1d0167d153b0fc6c9be56e265e653c3da7a#egg=InstagramAPI
ipfsapi==0.4.3
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
Kivy==1.10.1
Kivy-Garden==0.1.4
kivymd==0.1.2
macholib==1.11
moviepy==0.2.3.2
multidict==4.5.2
numpy==1.16.2
opencv-python==4.1.0.25
parso==0.4.0
pefile==2018.8.8
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plyvel==1.0.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycryptodome==3.8.1
pyenchant==2.0.0
Pygments==2.3.1
PyInstaller==3.4
python-dateutil==2.8.0
requests==2.11.1
requests-toolbelt==0.7.0
sanic==19.3.1
six==1.12.0
text-unidecode==1.2
tqdm==4.11.2
traitlets==4.3.2
typing-extensions==3.7.2
ujson==1.35
urllib3==1.24.1
uvloop==0.12.2
wcwidth==0.1.7
websockets==6.0
yarl==1.3.0
